requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
pydantic